Split file not to contain 2 namespaces
