+++ src/MessagePack.Experimental/FormatterHelpers.cs
@@ -363,44 +363,3 @@ namespace MessagePack.Experimental.Formatters
         }
     }
 }
-
-namespace MessagePack.Experimental.Resolvers
-{
-    internal static class PrimitiveArrayGetFormatterHelper
-    {
-        internal static object? GetFormatter(Type t)
-        {
-            if (!t.IsArray)
-            {
-                return null;
-            }
-
-            if (t == typeof(SByte[]))
-            {
-                return MessagePack.Experimental.Formatters.SByteArrayFormatter.Instance;
-            }
-
-            if (t == typeof(Int16[]))
-            {
-                return MessagePack.Experimental.Formatters.Int16ArrayFormatter.Instance;
-            }
-
-            if (t == typeof(Int32[]))
-            {
-                return MessagePack.Experimental.Formatters.Int32ArrayFormatter.Instance;
-            }
-
-            if (t == typeof(Single[]))
-            {
-                return MessagePack.Experimental.Formatters.SingleArrayFormatter.Instance;
-            }
-
-            if (t == typeof(Boolean[]))
-            {
-                return MessagePack.Experimental.Formatters.BooleanArrayFormatter.Instance;
-            }
-
-            return null;
-        }
-    }
-}
