+++ src/MessagePack.Experimental/PrimitiveArrayGetFormatterHelper.cs
@@ -1,0 +1,48 @@
+﻿// Copyright (c) All contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+/* THIS (.cs) FILE IS GENERATED. DO NOT CHANGE IT.
+ * CHANGE THE .tt FILE INSTEAD. */
+
+using System;
+
+namespace MessagePack.Experimental.Resolvers
+{
+    internal static class PrimitiveArrayGetFormatterHelper
+    {
+        internal static object? GetFormatter(Type t)
+        {
+            if (!t.IsArray)
+            {
+                return null;
+            }
+
+            if (t == typeof(SByte[]))
+            {
+                return Formatters.SByteArrayFormatter.Instance;
+            }
+
+            if (t == typeof(Int16[]))
+            {
+                return Formatters.Int16ArrayFormatter.Instance;
+            }
+
+            if (t == typeof(Int32[]))
+            {
+                return Formatters.Int32ArrayFormatter.Instance;
+            }
+
+            if (t == typeof(Single[]))
+            {
+                return Formatters.SingleArrayFormatter.Instance;
+            }
+
+            if (t == typeof(Boolean[]))
+            {
+                return Formatters.BooleanArrayFormatter.Instance;
+            }
+
+            return null;
+        }
+    }
+}
